+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Numeral system".
 - Tests for "Flood fill example".
 - Tests for "5D chests".
+- Tests for "Near-palindromes".
 
 ## [1.15.0] - 2022-11-30
 ### Added